libstand: Change sbrk() in libstand to take an intptr_t like libc's.

+++ lib/libstand/sbrk.c
@@ -24,7 +24,6 @@
  * SUCH DAMAGE.
  *
  * $FreeBSD: src/lib/libstand/sbrk.c,v 1.2.2.1 2000/05/18 08:00:57 ps Exp $
- * $DragonFly: src/lib/libstand/sbrk.c,v 1.2 2003/06/17 04:26:51 dillon Exp $
  */
 
 /*
@@ -53,7 +52,7 @@ getheap(size_t *sizep)
 }
 
 char *
-sbrk(int incr)
+sbrk(intptr_t incr)
 {
     char	*ret;
     
@@ -66,4 +65,3 @@ sbrk(int incr)
     errno = ENOMEM;
     return((char *)-1);
 }
-
